+++ web/app/business/hooks/numerical-guidance/indicator/use-indicator-list.hook.ts
