+++ web/app/business/hooks/numerical-guidance/indicator/use-live-indicators-value-view-model.hook.ts
@@ -1,10 +1,10 @@
 import { useMemo } from 'react';
-import { LiveIndicatorRequestParams } from '../../../store/querys/numerical-guidance/indicator.query';
-import { convertLiveIndicatorsValueViewModel } from '../../services/view-model/indicator-value/actual-indicators-value-view-model.service';
-import { useIndicatorBoardMetadataViewModel } from '../numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
-import { useIndicatorBoard } from '../numerical-guidance/indicator-board/use-indicator-board.hook';
+import { LiveIndicatorRequestParams } from '../../../../store/querys/numerical-guidance/indicator.query';
+import { convertLiveIndicatorsValueViewModel } from '../../../services/view-model/indicator-value/actual-indicators-value-view-model.service';
+import { useIndicatorBoardMetadataViewModel } from '../indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
+import { useIndicatorBoard } from '../indicator-board/use-indicator-board.hook';
 import { useIndicatorBoardMetadataStore } from '@/app/store/stores/numerical-guidance/indicator-board-metadata.store';
-import { useFetchLiveIndicatorsValueByType } from '../../../store/querys/numerical-guidance/indicator.query';
+import { useFetchLiveIndicatorsValueByType } from '../../../../store/querys/numerical-guidance/indicator.query';
 import { getStartDate } from '@/app/utils/date-formatter';
 
 export const useLiveIndicatorsValueViewModel = (indicatorBoardMetadataId?: string) => {
