+++ web/app/ui/components/numerical-guidance/custom-forecast-indicator/source-indicator-search-list.tsx
@@ -5,8 +5,8 @@ import TinyInput from '../../view/atom/tiny-input/tiny-input';
 import { SearchIcon } from '@heroicons/react/solid';
 import SourceIndicatorSearchListItem from './source-indicator-search-list-item';
 import { useState } from 'react';
-import { useIndicatorSearchList } from '@/app/business/hooks/indicator/use-indicator-search-list.hooks';
-import { useIndicatorListByType } from '@/app/business/hooks/indicator/use-indicator-list-by-type.hook';
+import { useIndicatorSearchList } from '@/app/business/hooks/numerical-guidance/indicator/use-indicator-search-list.hooks';
+import { useIndicatorListByType } from '@/app/business/hooks/numerical-guidance/indicator/use-indicator-list-by-type.hook';
 import { Indicator } from '@/app/business/services/view-model/indicator-list/indicators/indicator.service';
 import { IndicatorType } from '@/app/store/stores/numerical-guidance/indicator-list.store';
 import IndicatorTypeToggleGroup from '../../view/molecule/indicator-type-toggle-group';
